Adds module specific tests
